# docs/swagger.yaml
basePath: /
definitions:
  apiv1.UrlRequest:
    properties:
      destination:
        type: string
      label:
        type: string
    required:
    - destination
    - label
    type: object
  models.Url:
    properties:
      created_on:
        type: integer
      deleted_on:
        type: integer
      destination:
        type: string
      id:
        type: integer
      label:
        type: string
      modified_on:
        type: integer
    type: object
  requests.AuthRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
host: 127.0.0.1:8080
info:
  contact: {}
  description: This is documentation for redirect engine admin API
  title: URL management API
  version: "1.0"
paths:
  /api/v1/url:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Url'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get urls list
  /api/v1/url/:
    post:
      consumes:
      - application/json
      parameters:
      - description: URL label
        in: body
        name: url
        required: true
        schema:
          $ref: '#/definitions/apiv1.UrlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiv1.UrlRequest'
      security:
      - ApiKeyAuth: []
      summary: Create new url
  /api/v1/url/{id}:
    delete:
      parameters:
      - description: Url ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete an existing url
    get:
      consumes:
      - application/json
      parameters:
      - description: Url ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiv1.UrlRequest'
      security:
      - ApiKeyAuth: []
      summary: Get a single url
    put:
      consumes:
      - application/json
      parameters:
      - description: Url ID
        in: path
        name: id
        required: true
        type: integer
      - description: URL label
        in: body
        name: url
        required: true
        schema:
          $ref: '#/definitions/apiv1.UrlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiv1.UrlRequest'
      security:
      - ApiKeyAuth: []
      summary: Update an existing url
  /auth:
    post:
      parameters:
      - description: Auth data
        in: body
        name: url
        required: true
        schema:
          $ref: '#/definitions/requests.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"